# cmd/api/def/types.gen.go
// Package def provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.3 DO NOT EDIT.
package def

import (
	"time"

	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	APIKeyAuthScopes = "APIKeyAuth.Scopes"
	BearerAuthScopes = "BearerAuth.Scopes"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    *int    `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    *openapi_types.Email `json:"email,omitempty"`
	Password *string              `json:"password,omitempty"`
}

// RegisterationRequest defines model for RegisterationRequest.
type RegisterationRequest struct {
	ConfirmPassword *string              `json:"confirm_password,omitempty"`
	Email           *openapi_types.Email `json:"email,omitempty"`
	FirstName       *string              `json:"first_name,omitempty"`
	LastName        *string              `json:"last_name,omitempty"`
	Password        *string              `json:"password,omitempty"`
	TeamName        *string              `json:"team_name,omitempty"`
}

// RegisterationResponse defines model for RegisterationResponse.
type RegisterationResponse struct {
	Team *Team `json:"team,omitempty"`
	User *User `json:"user,omitempty"`
}

// Team defines model for Team.
type Team struct {
	CreatedAt *time.Time          `json:"created_at,omitempty"`
	Id        *openapi_types.UUID `json:"id,omitempty"`
	Name      *string             `json:"name,omitempty"`
	Slug      *string             `json:"slug,omitempty"`
	UpdatedAt *time.Time          `json:"updated_at,omitempty"`
}

// TokenResponse defines model for TokenResponse.
type TokenResponse struct {
	AccessToken  *string `json:"access_token,omitempty"`
	RefreshToken *string `json:"refresh_token,omitempty"`
}

// User defines model for User.
type User struct {
	CreatedAt  *time.Time           `json:"created_at,omitempty"`
	Email      *openapi_types.Email `json:"email,omitempty"`
	FirstName  *string              `json:"first_name,omitempty"`
	Id         *openapi_types.UUID  `json:"id,omitempty"`
	IsActive   *bool                `json:"is_active,omitempty"`
	IsVerified *bool                `json:"is_verified,omitempty"`
	LastName   *string              `json:"last_name,omitempty"`

	// TeamId A user can belong to many teams. This value represents the default team for the user.
	TeamId    *openapi_types.UUID `json:"team_id,omitempty"`
	UpdatedAt *time.Time          `json:"updated_at,omitempty"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = RegisterationRequest